// frontend/src/app/models/nekretnina.ts
export class Nekretnina{
    id: number;
    naziv: string;
    grad: string;
    opstina: string;
    ulica: string;
    broj: number;
    tip: string;
    sprat: number;
    zgrada: number;
    kvadratura: number;
    br_soba: string;
    namestena: number;
    galerija: Array<string>;
    namena: string;
    cena: number;
    vlasnik: string;
    odobrena: number;
    promovisana: number;
}